requirements.txt
mysql-connector-python
requests
ratelimit